requirements.txt
Django==3.1.14
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
psycopg2
PyJWT==1.7.1
pytz==2020.1
SQLAlchemy==1.4.36
sqlparse==0.3.1
